Split tests for php81
