+++ phpunit.xml.dist
@@ -18,9 +18,13 @@
     <testsuites>
         <testsuite name="Validator">
             <directory>./tests</directory>
+            <exclude>./tests/DataSet/PHP81</exclude>
+        </testsuite>
+
+        <testsuite name="Validator 81">
+            <directory phpVersion="8.1.0" phpVersionOperator=">=">./tests/DataSet/PHP81</directory>
         </testsuite>
     </testsuites>
-
     <coverage>
         <include>
             <directory>./src</directory>
