+++ tests/DataSet/AttributeDataSet80Test.php
@@ -1,0 +1,114 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Validator\Tests\DataSet;
+
+use PHPUnit\Framework\TestCase;
+use Yiisoft\Validator\Attribute\Embedded;
+use Yiisoft\Validator\DataSet\AttributeDataSet;
+use Yiisoft\Validator\Rule\Count;
+use Yiisoft\Validator\Rule\Each;
+use Yiisoft\Validator\Rule\GroupRule;
+use Yiisoft\Validator\Rule\HasLength;
+use Yiisoft\Validator\Rule\Nested;
+use Yiisoft\Validator\Rule\Number;
+use Yiisoft\Validator\Rule\Required;
+use Yiisoft\Validator\RuleInterface;
+use Yiisoft\Validator\Tests\Data\Charts\Chart;
+use Yiisoft\Validator\Tests\Data\Charts\Coordinate;
+use Yiisoft\Validator\Tests\Data\TitleTrait;
+use Yiisoft\Validator\Tests\Stub\NotRuleAttribute;
+
+final class AttributeDataSet80Test extends TestCase
+{
+    /**
+     * @dataProvider dataProvider
+     *
+     * @param RuleInterface[]|RuleInterface[][]|RuleInterface[][][] $expectedRules
+     */
+    public function testCollectRules(object $object, array $expectedRules): void
+    {
+        $dataSet = new AttributeDataSet($object);
+
+        $this->assertEquals($expectedRules, $dataSet->getRules());
+    }
+
+    public function dataProvider(): array
+    {
+        return [
+            [
+                new class () {
+                },
+                [],
+            ],
+            [
+                new class () {
+                    private $property1;
+                },
+                [],
+            ],
+            [
+                new class () {
+                    #[NotRuleAttribute]
+                    private $property1;
+                },
+                [],
+            ],
+            [
+                new class () {
+                    #[Required(skipOnEmpty: true)]
+                    private $property1;
+                },
+                [
+                    'property1' => [
+                        new Required(skipOnEmpty: true),
+                    ],
+                ],
+            ],
+            [
+                new class () {
+                    use TitleTrait;
+                },
+                [
+                    'title' => [
+                        new HasLength(max: 255),
+                    ],
+                ],
+            ],
+            [
+                new class () {
+                    #[Required(skipOnEmpty: true)]
+                    #[HasLength(skipOnEmpty: true)]
+                    private $property1;
+                    #[Required(skipOnEmpty: true)]
+                    #[HasLength(skipOnEmpty: true)]
+                    private $property2;
+                },
+                [
+                    'property1' => [
+                        new Required(skipOnEmpty: true),
+                        new HasLength(skipOnEmpty: true),
+                    ],
+                    'property2' => [
+                        new Required(skipOnEmpty: true),
+                        new HasLength(skipOnEmpty: true),
+                    ],
+                ],
+            ],
+            [
+                new class () {
+                    #[HasLength(skipOnEmpty: true)]
+                    #[HasLength(skipOnEmpty: false)]
+                    private $property1;
+                },
+                [
+                    'property1' => [
+                        new HasLength(skipOnEmpty: true),
+                        new HasLength(skipOnEmpty: false),
+                    ],
+                ],
+            ],
+        ];
+    }
+}
